requirements.txt
Pillow~=9.4.0
toml~=0.10.2
